+++ src/paco/models/schemas.py
@@ -1529,7 +1529,7 @@ This is a base schema which defines metadata useful to categorize an alarm.
     )
     notification_groups = schema.List(
         readonly = True,
-        title="List of notificationn groups the alarm is subscribed to.",
+        title="List of notification groups the alarm is subscribed to.",
         value_type=schema.TextLine(title="Notification group name"),
         required=False,
     )
@@ -4066,28 +4066,35 @@ class IBlockDeviceMapping(IParent):
         required=False
     )
 
-class IASGRollingUpdatePolicy(INamed, IDeployable):
+class IASGRollingUpdatePolicy(INamed):
     """
-Auto Scaling Group Roling Update Policy
+AutoScalingRollingUpdate Policy
     """
+    enabled = schema.Bool(
+        title="Enable an UpdatePolicy for the ASG",
+        description="",
+        default=True,
+        required=False,
+    )
     max_batch_size = schema.Int(
         title="Maximum batch size",
         description="",
         default=1,
+        min=1,
         required=False,
     )
     min_instances_in_service = schema.Int(
         title="Minimum instances in service",
         description="",
-        default=1,
+        default=0,
+        min=0,
         required=False,
     )
     pause_time = schema.TextLine(
         title="Minimum instances in service",
-        description="Healthy success timeout",
+        description="Must be in the format PT#H#M#S",
         required=False,
-        default='PT0S'
-        #constraint=IsValidUpdatePolicyPauseTime
+        default='',
     )
     wait_on_resource_signals = schema.Bool(
         title="Wait for resource signals",
@@ -4097,7 +4104,7 @@ Auto Scaling Group Roling Update Policy
 
 class IASG(IResource, IMonitorable):
     """
-An Auto Scaling Group (ASG) contains a collection of Amazon EC2 instances that are treated as a
+An AutoScalingGroup (ASG) contains a collection of Amazon EC2 instances that are treated as a
 logical grouping for the purposes of automatic scaling and management.
 
 The Paco ASG resource provisions an AutoScalingGroup as well as LaunchConfiguration and TargetGroups
@@ -4135,7 +4142,6 @@ for that ASG.
     that will install a CloudWatch Agent and configure it to collect all specified metrics and log sources.
 
     ``secrets``: Adds a policy to the Instance Role which allows instances to access the specified secrets.
-
 
 .. code-block:: yaml
     :caption: example ASG configuration
@@ -4175,8 +4181,13 @@ for that ASG.
     instance_monitoring: true
     instance_type: t2.medium
     desired_capacity: 1
-    max_instances: 1
+    max_instances: 3
     min_instances: 1
+    rolling_update_policy:
+      max_batch_size: 1
+      min_instances_in_service: 1
+      pause_time: PT3M
+      wait_on_resource_signals: false
     target_groups:
       - paco.ref netenv.mynet.applications.app.groups.web.resources.alb.target_groups.cloud
     security_groups:
@@ -4184,8 +4195,6 @@ for that ASG.
     segment: private
     termination_policies:
       - Default
-    update_policy_max_batch_size: 1
-    update_policy_min_instances_in_service: 0
     scaling_policy_cpu_average: 60
     launch_options:
         cfn_init_config_sets:
@@ -4238,14 +4247,95 @@ for that ASG.
     user_data_script: |
       echo "Hello World!"
 
+
+AutoScalingGroup Rolling Update Policy
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+When changes are applied to an AutoScalingGroup that modify the configuration of newly launched instances,
+AWS can automatically launch instances with the new configuration and terminate old instances that have stale configuration.
+This can be configured so that there is no interruption of service as the new instances gradually replace old ones.
+This configuration is set with the ``rolling_update_policy`` field.
+
+The rolling update policy must be able to work within the minimum/maximum number of instances in the ASG.
+Consider the following ASG configuration.
+
+.. code-block:: yaml
+    :caption: example ASG configuration
+
+    type: ASG
+    max_instances: 2
+    min_instances: 1
+    desired_capacity: 1
+    rolling_update_policy:
+      max_batch_size: 1
+      min_instances_in_service: 1
+      pause_time: PT0S # default setting
+      wait_on_resource_signals: false # default setting
+
+This will normally run a single instance in the ASG. The ASG is never allowed to launch more than 2 instances at one time.
+When an update happens, a new batch of instances is launched - in this example just one instance. There wil be only 1 instance
+in service, but the capacity will be at 2 instances will the new instance is launched. After the instance
+is put into service by the ASG, it will immediately terminate the old instance.
+
+The ``wait_on_resource_signals`` can be set to tell AWS CloudFormation to wait on making changes to the AutoScalingGroup configuration
+until a new instance is finished configuring and installing applications and is ready for service. If this field is enabled,
+then the ``pause_time`` default is PT05 (5 minutes). If CloudFormation does not get a SUCCESS signal within the ``pause_time``
+then it will mark the new instance as failed and terminate it.
+
+If you use ``pause_time`` with the default ``wait_on_resource_signals: false`` then AWS will simply wait for the full
+duration of the pause time and then consider the instance ready. ``pause_time`` is in format PT#H#M#S, where each # is the number of
+hours, minutes, and seconds, respectively. The maximum ``pause_time`` is one hour. For example:
+
+.. code-block:: yaml
+
+    pause_time: PT0S # 0 seconds
+    pause_time: PT5M # 5 minutes
+    pause_time: PT2M30S # 2 minutes and 30 seconds
+
+ASGs will use default settings for a rolling update policy. If you do not want to use an update policies at all, then
+you must disable the ``rolling_update_policy`` explicitly:
+
+.. code-block:: yaml
+
+    type: ASG
+    rolling_update_policy:
+      enabled: false
+
+With no rolling update policy, when you make configuration changes, then existing instances with old configuration will
+continue to run and instances with the new configuration will not happen until the AutoScalingGroup needs to launch new
+instances. You must be careful with this approach as you can not know 100% that your new configuration launches instances
+proprely until some point in the future when new instances are requested by the ASG.
+
+.. sidebar:: Prescribed Automation
+
+    Paco can help you send signals to CloudFormation when using ``wait_on_resource_signals``.
+    If you set ``wait_on_resource_signals: true`` then Paco will automatically grant the needed ``cloudformation:SignalResource`` and
+    ``cloudformation:DescribeStacks`` to the IAM Role associated with the instance for you. Paco also provides an
+    ``ec2lm_signal_asg_resource`` BASH function available in your ``user_data_script`` that you can run to signal the instance is
+    ready: ``ec2lm_signal_asg_resource SUCCESS`` or ``ec2lm_signal_asg_resource SUCCESS``.
+
+    If you want to wait until load balancer health checks are passing before an instance is considered healthy, then send the SUCCESS
+    signal to CloudFormation, you will need to configure this yourself.
+
+        .. code-block:: bash
+            :caption: example ASG signalling using ELB health checks
+
+            'until [ "$state" == "\"InService\"" ]; do state=$(aws --region ${AWS::Region} elb describe-instance-health
+            --load-balancer-name ${ElasticLoadBalancer}
+            --instances $(curl -s http://169.254.169.254/latest/meta-data/instance-id)
+            --query InstanceStates[0].State); sleep 10; done'
+
+
+See the AWS documentation for more information on how `AutoScalingRollingUpdate Policy`_ configuration is used.
+
+.. _AutoScalingRollingUpdate Policy: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-attribute-updatepolicy.html#cfn-attributes-updatepolicy-replacingupdate
+
     """
     @invariant
     def min_instances(obj):
         if obj.rolling_update_policy != None:
             if obj.rolling_update_policy != None and obj.rolling_update_policy.min_instances_in_service >= obj.max_instances:
                 raise Invalid("ASG rolling_update_policy.min_instances_in_service must be less than max_instances.")
-        elif obj.update_policy_min_instances_in_service >= obj.max_instances:
-            raise Invalid("ASG update_policy_min_instances_in_service must be less than max_instances.")
         if obj.min_instances > obj.max_instances:
             raise Invalid("ASG min_instances must be less than or equal to max_instances.")
         if obj.desired_capacity > obj.max_instances:
@@ -4400,18 +4490,6 @@ for that ASG.
         default=1,
         required=False,
     )
-    update_policy_max_batch_size = schema.Int(
-        title="Update policy maximum batch size",
-        description="",
-        default=0,
-        required=False,
-    )
-    update_policy_min_instances_in_service = schema.Int(
-        title="Update policy minimum instances in service",
-        description="",
-        default=0,
-        required=False,
-    )
     scaling_policies = schema.Object(
         title='Scaling Policies',
         schema=IASGScalingPolicies,
@@ -4481,8 +4559,7 @@ for that ASG.
         title="Rolling Update Policy",
         description="",
         schema=IASGRollingUpdatePolicy,
-        default=None,
-        required=False
+        required=True
     )