+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 Changelog for paco.models
 =========================
 
-7.0.3 (unreleased)
+8.0.0 (unreleased)
 ------------------
 
-- Nothing changed yet.
+### Migration
+
+- ASG field's ``update_policy_max_batch_size`` and ``update_policy_min_instances_in_service`` are removed.
+  Instead use the ASG field ``rolling_update_policy`` and set ``max_batch_size`` and ``min_instances_in_service``.
 
 
 7.0.2 (2020-03-14)